[WIP] refactoring
