+++ pgdrive/scene_creator/blocks/ramp.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from pgdrive.pg_config.parameter_space import Parameter, BlockParameterSpace
 from pgdrive.pg_config.pg_space import PgSpace
-from pgdrive.scene_creator.basic_utils import check_lane_on_road, ExtendStraightLane, CreateRoadFrom, Decoration, \
+from pgdrive.scene_creator import check_lane_on_road, ExtendStraightLane, CreateRoadFrom, Decoration, \
     sharpbend, CreateAdverseRoad
 from pgdrive.scene_creator.blocks.block import Block, BlockSocket
 from pgdrive.scene_creator.lanes.lane import LineType
