+++ pgdrive/scene_creator/road/road_network.py
@@ -3,7 +3,7 @@ import logging
 from typing import List, Tuple, Dict
 
 import numpy as np
-
+from pgdrive.scene_creator.basic_utils import Decoration
 from pgdrive.scene_creator.lanes.lane import LineType, AbstractLane
 from pgdrive.scene_creator.lanes.straight_lane import StraightLane
 from pgdrive.scene_creator.road.road import Road
@@ -30,7 +30,6 @@ class RoadNetwork:
         return ret
 
     def __iadd__(self, other):
-        from pgdrive.scene_creator.basic_utils import Decoration
         set_1 = set(self.graph) - {Decoration.start, Decoration.end}
         set_2 = set(other.graph) - {Decoration.start, Decoration.end}
         if len(set_1.intersection(set_2)) == 0:
@@ -43,7 +42,6 @@ class RoadNetwork:
             raise ValueError("Same start node in two road network")
 
     def __isub__(self, other):
-        from pgdrive.scene_creator.basic_utils import Decoration
         intersection = self.graph.keys() & other.graph.keys() - {Decoration.start, Decoration.end}
         if len(intersection) != 0:
             for k in intersection:
@@ -61,7 +59,6 @@ class RoadNetwork:
         return ret
 
     def get_all_decoration_lanes(self) -> List:
-        from pgdrive.scene_creator.basic_utils import Decoration
         if Decoration.start in self.graph:
             return self.graph[Decoration.start][Decoration.end]
         else:
@@ -70,7 +67,6 @@ class RoadNetwork:
     def update_decoration_lanes(self, lanes):
         if len(lanes) == 0:
             return
-        from pgdrive.scene_creator.basic_utils import Decoration
         if Decoration.start in self.graph:
             self.graph.pop(Decoration.start, None)
         self.graph[Decoration.start] = {Decoration.end: lanes}
@@ -103,7 +99,6 @@ class RoadNetwork:
         return ret
 
     def remove_road(self, road):
-        from pgdrive.scene_creator.road.road import Road
         assert isinstance(road, Road), "Only Road Type can be deleted"
         ret = self.graph[road.start_node].pop(road.end_node)
         if len(self.graph[road.start_node]) == 0:
@@ -111,7 +106,6 @@ class RoadNetwork:
         return ret
 
     def add_road(self, road, lanes: List):
-        from pgdrive.scene_creator.road.road import Road
         assert isinstance(road, Road), "Only Road Type can be added to road network"
         if road.start_node not in self.graph:
             self.graph[road.start_node] = {}
