+++ pgdrive/world/__init__.py
@@ -1,1 +1,4 @@
-from pgdrive.world.constants import *
+from pgdrive.world.constants import (
+    COLOR, COLLISION_INFO_COLOR, RENDER_MODE_OFFSCREEN, RENDER_MODE_ONSCREEN, RENDER_MODE_NONE, PG_EDITION,
+    HELP_MESSAGE, DEBUG_MESSAGE
+)
