+++ pgdrive/world/onscreen_message.py
@@ -2,7 +2,6 @@ from typing import Optional, Union
 
 from direct.showbase import OnScreenDebug
 from panda3d.core import Vec4
-
 from pgdrive.world.constants import HELP_MESSAGE, DEBUG_MESSAGE
 
 
@@ -12,10 +11,9 @@ class PgOnScreenMessage(OnScreenDebug.OnScreenDebug):
     """
     POS = (0.1, -0.2)
 
-    def __init__(self, refresh_plain_text=False):
+    def __init__(self, refresh_plain_text=False, debug=False):
         super(PgOnScreenMessage, self).__init__()
-        from pgdrive.world.pg_world import PgWorld
-        self.debug = PgWorld.DEBUG
+        self.debug = debug
         self.enabled = True
         self.load()
         self.plain_text = set()
